requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
numpy==1.26.4
scipy==1.13.0
tensorflow-cpu==2.16.1
scikit-learn==1.5.0
pydantic==2.7.1
opencv-python-headless==4.9.0.80
pandas==2.2.0
